refactor: useContract 생성 후 contract/[id] 페이지에서 사용
